对back文件夹代码和idea进行了一点规范化

+++ src/main/java/cn/bb/sourceideamanage/service/back/impl/BackIdeaServiceImpl.java
@@ -6,6 +6,8 @@ import cn.bb.sourceideamanage.entity.Tag;
 import cn.bb.sourceideamanage.service.back.BackIdeaService;
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import lombok.extern.slf4j.Slf4j;
+import net.sf.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.Cacheable;
@@ -16,13 +18,16 @@ import javax.annotation.Resource;
 import java.util.List;
 
 @Service
+@Slf4j
 public class BackIdeaServiceImpl implements BackIdeaService {
 
     @Resource
     BackIdeaMapper mapper;
 
+    @Autowired
+    JSONObject jsonObject;
+
     @Override
-    @Cacheable(cacheNames = "backIdeasPage",key = "'backIdeasPage=[page='+#page+'][size='+#size+'][ideaName=['+#ideaName+'][tagName'+#tagName+']'")
     public PageInfo<BackIdea> findBackIdeaByPage(int page, int size, String ideaName, String tagName) {
         PageHelper.startPage(page,size);
         List<BackIdea> ideas = mapper.findAllBackIdea(ideaName, tagName);
@@ -37,7 +42,20 @@ public class BackIdeaServiceImpl implements BackIdeaService {
 
     @Override
     @Transactional(rollbackFor = Exception.class)
-    public void deleteIdea(Integer ideaId) {
-        mapper.deleteIdea(ideaId);
+    public String deleteIdea(Integer ideaId) {
+        if(ideaId == null){
+            jsonObject.put("msg","删除失败!");
+            jsonObject.put("success","0");
+            log.error("BackIdeaService 删除错误");
+        }else{
+            //删除想法
+            mapper.deleteIdea(ideaId);
+            //删除想法评论
+            mapper.deleteIdeaComment(ideaId);
+            jsonObject.put("msg","删除成功!");
+            jsonObject.put("success","1");
+            log.info("BackIdeaService删除成功!");
+        }
+        return jsonObject.toString();
     }
 }
